Fix compiler warning (0014)

+++ xbmc/games/controllers/windows/GUIControllerList.cpp
@@ -157,7 +157,7 @@ void CGUIControllerList::OnSelect(unsigned int controllerIndex)
 
 void CGUIControllerList::ResetController(void)
 {
-  if (0 <= m_focusedController && m_focusedController < m_controllers.size())
+  if (0 <= m_focusedController && m_focusedController < (int)m_controllers.size())
   {
     const std::string strControllerId = m_controllers[m_focusedController]->ID();
     g_peripherals.ResetButtonMaps(strControllerId);
